master

+++ db/migrate/20200709113339_add_norm_to_exam_unit.rb
@@ -1,6 +1,6 @@
 class AddNormToExamUnit < ActiveRecord::Migration[5.2]
   def change
-    add_column :exam_units, :norm_min, :integer
-    add_column :exam_units, :norm_max, :integer
+    add_column :exam_units, :norm_min, :decimal
+    add_column :exam_units, :norm_max, :decimal
   end
 end
