+++ db/migrate/20200709115148_create_order_item_results.rb
@@ -2,7 +2,7 @@ class CreateOrderItemResults < ActiveRecord::Migration[5.2]
   def change
     create_table :order_item_results do |t|
       t.references :order_item, foreign_key: true
-      t.integer :result
+      t.decimal :result
       t.datetime :result_time
 
       t.timestamps
